requirements.txt
requests
pytest
python-dotenv
pytest-html
uuid